% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSC.R
\name{calc_WSC}
\alias{calc_WSC}
\title{Water Stress Coefficient (WSC)}
\usage{
calc_WSC(
  wsc_method,
  D_e = NULL,
  SWIR1 = NULL,
  SWIR2 = NULL,
  NIR = NULL,
  prec = NULL
)
}
\arguments{
\item{wsc_method}{The method to be used for WSC calculation. Different
  approaches require different input data. The following approaches are
  implemented:
\describe{

\item{deficit}{Requires the vapour pressure deficit \strong{D_e} in kPa. \cr
  \eqn{WSC = 1.2 \cdot exp(-0.35 \cdot D_e) - 0.2}. \cr Tang et al. (2014)}

\item{SIMI}{Remote sensinsg data driven. Requires bands \strong{SWIR1}
(1.628 - 1.652 µm) and \strong{SWIR2} (2.105 - 2.155 µm) to be provided.
  WSC is calculated per pixel. The min and max of SIMI are also calculated
  per pixel. \cr
  \eqn{SIMI = 0.7071 \cdot sqrt(SWIR1^2 + SWIR2^2)} \cr
  \eqn{NSIMI = (SIMI - min(SIMI)) / (max(SIMI) - min(SIMI))} \cr
  \eqn{WSC = 0.5 + 0.5 * NSIMI} \cr
  Wu et al. (2022)}

\item{LSWI}{Remote sensinsg and meteorological data driven. Requires bands
\strong{SWIR1} (1.628 - 1.652 µm) and \strong{NIR} to be provided, as well as \strong{prec}. \cr
WSC is calculated per pixel. The max of LSWI is also calculated per pixel.
A rolling sum of the last 30 days is applied to prec to get the monthly
precipitation. \cr
  \eqn{LSWI = (NIR - SWIR1) / (NIR + SWIR1)} \cr
  \eqn{WSC_{vpm,in} = (1 - (1 + LSWI) / (1 + LSWI_{max})) + 0.5} \cr
  \eqn{WSC = prec / prec_{max} \cdot WSC_{vpm,in} + 0.5} \cr
  Bao et al. (2015)}

}}

\item{D_e}{vector of mean daily vapour pressure deficit in kPa}

\item{SWIR1}{A raster stack where each layer is one time point.}

\item{SWIR2}{A raster stack where each layer is one time point.}

\item{NIR}{A raster stack where each layer is one time point.}

\item{prec}{vector of daily precipitation in mm}
}
\value{
A vector or RasterStack of the water stress coefficient with one
  entry (layer) per day.
}
\description{
This function calculates the Water Stress Coefficient (WSC) for
  the CASA model.
}
\references{
\insertRef{Tang2014}{remotepp} \cr
\insertRef{Bao2016}{remotepp} \cr
\insertRef{Wu2022}{remotepp} \cr
}
