% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gif.R
\name{create_gif}
\alias{create_gif}
\title{Create a gif from a time series of raster files}
\usage{
create_gif(input_files, plot_FUN, out_file)
}
\arguments{
\item{input_files}{character list of files}

\item{plot_FUN}{function used to plot the data}

\item{out_file}{character path to the output file to be generated.}
}
\description{
Create an animated gif file of a time series of raster data.
}
