% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_per_photon.R
\name{energy_per_photon}
\alias{energy_per_photon}
\title{Convert PAR from mol to J}
\usage{
energy_per_photon(x, wavelength)
}
\arguments{
\item{x}{photons in mol}

\item{wavelength}{wavelength in nm}
}
\value{
energy in J
}
\description{
Convert PAR from mol to J
}
