% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_FAO56.R
\name{ET_FAO56}
\alias{ET_FAO56}
\title{Calculate Evapotranspiration using Penman-Monteith FAO56}
\usage{
ET_FAO56(irradiance, temperature, ea, u_2, pressure, ...)
}
\arguments{
\item{irradiance}{at crop surface in MJ m^-2 day^-1}

\item{temperature}{mean daily temperature in °C}

\item{ea}{vapour pressure in kPa}

\item{u_2}{wind speed at 2 m in m s^-1}

\item{pressure}{air pressure in kPa}

\item{G}{soil heat flux in MJ m^-2 day^-1. Default is 0}
}
\value{
reference evapotranspiration in mm m^-2 d^-1
}
\description{
Calculate Evapotranspiration using Penman-Monteith FAO56
}
\references{
ISBN 92-5-104219-5
}
