% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_raster.R
\name{interpolate_raster}
\alias{interpolate_raster}
\title{interpolate raster}
\usage{
interpolate_raster(
  input_files,
  in_dates,
  out_dates,
  out_dir,
  interpolation_function = function(xin, y, xout) (approx(xin, y, xout, ties = mean))$y,
  interpolation_arguments = list(),
  overwrite = FALSE,
  cores = max(c(parallelly::availableCores() - 1, 1))
)
}
\description{
Interpolate a raster timeseries from file to file.
}
