% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_d2d.R
\name{map_d2d}
\alias{map_d2d}
\title{Apply a function disk to disk}
\usage{
map_d2d(input_files, read_FUN, change_FUN, save_FUN, out_dir, overwrite, ...)
}
\arguments{
\item{input_files}{character list of files}

\item{read_FUN}{function to read in a file from `input_files`}

\item{save_FUN}{function to save the object. Must match the format `f(object, path)`}

\item{out_dir}{character path where to store the output}

\item{overwrite}{logical}

\item{...}{additional arguments passed to FUN}

\item{FUN}{function to apply to each file in `input_files`}
}
\description{
Apply a function to a list of files and store again on disk.
Use parallel processing by running `future::plan(future::sequential)` beforehand.
}
