% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{ndvi}
\alias{ndvi}
\title{NDVI}
\usage{
ndvi(input_raster, red_band, ni_band)
}
\arguments{
\item{input_raster}{A raster stack or the path to a raster file.}

\item{red_band}{integer indicating the layer of `input_raster` that holds the
red band}

\item{ni_band}{integer indicating the layer of `input_raster` that holds the
near infrared band}
}
\description{
Calculate the normalised difference vegetation index of a raster layer.
}
