% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPFD_to_irradiance.R
\name{PPFD_to_irradiance}
\alias{PPFD_to_irradiance}
\title{Convert PAR}
\usage{
PPFD_to_irradiance(x)
}
\arguments{
\item{x}{PAR in µmol s^-1 (or µmol s^-1 m^-2; ...)}
}
\value{
\eqn{R_S} in W (or W m^-2; J s^-1 m^-2; ...). This is calculated
  using a simple factor of 4.57.
  nm, the range of PAR.  \cr \cr \eqn{R_S {[}\frac{J}{s}{]} =
  \frac{1}{4.57}{[}\frac{J}{µmol}{]} \cdot PPFD {[}\frac{µmol}{s}{]}}
}
\description{
Convert PAR
}
\references{
\insertRef{Reis2020}{remotepp} \cr
}
