% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpp_CASA.R
\name{rpp_CASA}
\alias{rpp_CASA}
\title{rpp_CASA}
\usage{
rpp_CASA(
  NPP = terra::rast(),
  PAR = c(),
  FAPAR = terra::rast(),
  TSC1 = c(),
  TSC2 = c(),
  WSC = c(),
  epsilon_max = numeric()
)
}
\arguments{
\item{NPP}{`terra::rast()` of the NPP in g-C m^-2 d^-1 with each layer representing one day.}

\item{PAR}{Incoming photosyntheticly active radiation (PAR) in MJ m^-2 d^-1}

\item{FAPAR}{Fraction of absorbed PAR (0,1) as a raster stack with each layer
representing one day. If missing, is calculated using `calc_FAPAR()`}

\item{TSC1}{Temperature stress coefficient 1. If missing, is calculated using `calc_TSC1()`}

\item{TSC2}{Temperature stress coefficient 2. If missing, is calculated using `calc_TSC2()`}

\item{WSC}{Water stress coefficient. If missing, is calculated using `calc_WSC()`}

\item{epsilon_max}{The maximum light use efficiency (LUE) in g-C MJ^{-1}}

\item{x}{a CASA model list}
}
\description{
The class rpp_CASA is the output from running the `CASA()` model. It is a named list that includes the
output layer NPP, as well as all the other parameters derived from the input.
}
