% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpp_MODIS.R
\name{rpp_MODIS}
\alias{rpp_MODIS}
\title{rpp_MODIS}
\usage{
rpp_MODIS(
  GPP = terra::rast(),
  PAR = c(),
  FAPAR = terra::rast(),
  TMIN_scalar = c(),
  VPD_scalar = c(),
  epsilon_max = numeric(),
  dates = c()
)
}
\arguments{
\item{GPP}{`terra::rast()` of the GPP in g-C m^-2 d^-1 with each layer representing one day.}

\item{PAR}{Incoming photosyntheticly active radiation (PAR) in MJ m^-2 d^-1}

\item{FAPAR}{Fraction of absorbed PAR (0,1) as a raster stack with each layer
representing one day. If missing, is calculated using `calc_FAPAR()`}

\item{TMIN_scalar}{Temperature stress factor, calculated by `calc_TMIN_scalar()`.}

\item{VPD_scalar}{Vapor pressure deficit factor, calculated by `calc_VPD_scalar()` .}

\item{epsilon_max}{The maximum light use efficiency (LUE) in g-C MJ^{-1}}

\item{dates}{A vector the same length as PAR giving the dates.}

\item{x}{a MODIS model list}
}
\description{
The class rpp_MODIS is the output from running the `MODIS()` model. It is a named list that includes the
output layer GPP, as well as all the other parameters derived from the input.
}
