% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCASA.R
\name{saveCASA}
\alias{saveCASA}
\title{Save and load a rpp_CASA model result}
\usage{
saveCASA(x, file_path, overwrite = FALSE)
}
\arguments{
\item{x}{A `CASA()` model result of class rpp_CASA}
}
\description{
Save a CASA model result to disk. This saves the object as an
RDS and the raster layers as tifs to a specified folder.
}
