% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveMODIS.R
\name{saveMODIS}
\alias{saveMODIS}
\title{Save and load a rpp_MODIS model result}
\usage{
saveMODIS(x, file_path, overwrite = FALSE)
}
\arguments{
\item{x}{A `MODIS()` model result of class rpp_MODIS}
}
\description{
Save a MODIS model result to disk. This saves the object as an
RDS and the raster layers as tifs to a specified folder.
}
