% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_TMIN_scalar.R
\name{calc_TMIN_scalar}
\alias{calc_TMIN_scalar}
\title{MODIS: TMIN_scalar}
\usage{
calc_TMIN_scalar(temperature, TMIN_min = -8, TMIN_max = 12.02)
}
\arguments{
\item{temperature}{vector of mean daily temperature in °C}

\item{TMIN_min}{daily minimum temperature at which epsilon = epsilon_max}

\item{TMIN_max}{daily minimum temperature at which epsilon = 0}
}
\description{
This function calculates the TMIN_scalar coefficient for the MODIS GPP model.
}
\references{
\insertRef{Running2015}{remotepp} \cr
}
