% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSC.R
\name{calc_TSC1}
\alias{calc_TSC1}
\alias{calc_TSC2}
\title{Temperature Stress Coefficient (TSC)}
\usage{
calc_TSC1(temperature, temperature_opt)

calc_TSC2(temperature, temperature_opt)
}
\arguments{
\item{temperature}{vector of mean daily temperature in °C}

\item{temperature_opt}{the mean temperature of the month, where NDVI is highest in °C}
}
\description{
This function calculates the temperature Stress Coefficients (TSC) for
  the CASA model.
}
\references{
\insertRef{Potter1993}{remotepp} \cr
}
