% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_VPD_scalar.R
\name{calc_VPD_scalar}
\alias{calc_VPD_scalar}
\title{MODIS: VPD_scalar}
\usage{
calc_VPD_scalar(D_e, VPD_min = -650, VPD_max = 5300)
}
\arguments{
\item{D_e}{vector of mean daily vapor pressure deficit in kPa}

\item{VPD_min}{The daylight average vapor pressure deficit at which epsilon = epsilon_max [Pa]}

\item{VPD_max}{The daylight average vapor pressure deficit at which at which epsilon = 0 [Pa]}
}
\description{
This function calculates the VPD_scalar coefficient for the MODIS GPP model.
}
\references{
\insertRef{Running2015}{remotepp} \cr
}
