% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAPAR.R
\name{calc_FAPAR}
\alias{calc_FAPAR}
\title{FAPAR: Fraction of absorbed PAR}
\usage{
calc_FAPAR(NDVI, FAPAR_max = 0.95, FAPAR_min = 0.001, cores = 1)
}
\arguments{
\item{NDVI}{A raster(stack) of NDVI values}

\item{FAPAR_max}{constant, see Wang et al. (2017)}

\item{FAPAR_min}{constant, see Wang et al. (2017)}
}
\description{
FAPAR: Fraction of absorbed PAR
}
\references{
\insertRef{Wang2017}{remotepp} \cr
}
