% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASA.R
\name{CASA}
\alias{CASA}
\title{The CASA model}
\usage{
CASA(
  PAR,
  FAPAR = NULL,
  WSC = NULL,
  TSC1 = NULL,
  TSC2 = NULL,
  epsilon_max = 0.608,
  ...
)
}
\arguments{
\item{PAR}{Incoming photosyntheticly active radiation (PAR) in MJ m^-2 d^-1}

\item{FAPAR}{Fraction of absorbed PAR (0,1) as a raster stack with each layer
representing one day. If missing, is calculated using `calc_FAPAR()`}

\item{WSC}{Water stress coefficient. If missing, is calculated using `calc_WSC()`}

\item{TSC1}{Temperature stress coefficient 1. If missing, is calculated using `calc_TSC1()`}

\item{TSC2}{Temperature stress coefficient 2. If missing, is calculated using `calc_TSC2()`}

\item{epsilon_max}{The maximum light use efficiency (LUE) in g-C MJ^{-1}}

\item{...}{
  Arguments passed on to \code{\link[=calc_WSC]{calc_WSC}}, \code{\link[=calc_TSC1]{calc_TSC1}}, \code{\link[=calc_TSC2]{calc_TSC2}}, \code{\link[=calc_FAPAR]{calc_FAPAR}}
  \describe{
    \item{\code{wsc_method}}{The method to be used for WSC calculation. Different
  approaches require different input data. The following approaches are
  implemented:
\describe{

\item{deficit}{Requires the vapour pressure deficit \strong{D_e} in kPa. \cr
  \eqn{WSC = 1.2 \cdot exp(-0.35 \cdot D_e) - 0.2}. \cr Tang et al. (2014)}

\item{SIMI}{Remote sensinsg data driven. Requires bands \strong{SWIR1}
(1.628 - 1.652 µm) and \strong{SWIR2} (2.105 - 2.155 µm) to be provided.
  WSC is calculated per pixel. The min and max of SIMI are also calculated
  per pixel. \cr
  \eqn{SIMI = 0.7071 \cdot sqrt(SWIR1^2 + SWIR2^2)} \cr
  \eqn{NSIMI = (SIMI - min(SIMI)) / (max(SIMI) - min(SIMI))} \cr
  \eqn{WSC = 0.5 + 0.5 * NSIMI} \cr
  Wu et al. (2022)}

\item{LSWI}{Remote sensinsg and meteorological data driven. Requires bands
\strong{SWIR1} (1.628 - 1.652 µm) and \strong{NIR} to be provided, as well as \strong{prec}. \cr
WSC is calculated per pixel. The max of LSWI is also calculated per pixel.
A rolling sum of the last 30 days is applied to prec to get the monthly
precipitation. \cr
  \eqn{LSWI = (NIR - SWIR1) / (NIR + SWIR1)} \cr
  \eqn{WSC_{vpm,in} = (1 - (1 + LSWI) / (1 + LSWI_{max})) + 0.5} \cr
  \eqn{WSC = prec / prec_{max} \cdot WSC_{vpm,in} + 0.5} \cr
  Bao et al. (2015)}

}}
    \item{\code{D_e}}{vector of mean daily vapour pressure deficit in kPa}
    \item{\code{SWIR1}}{A raster stack where each layer is one time point.}
    \item{\code{SWIR2}}{A raster stack where each layer is one time point.}
    \item{\code{NIR}}{A raster stack where each layer is one time point.}
    \item{\code{prec}}{vector of daily precipitation in mm}
    \item{\code{temperature}}{vector of mean daily temperature in °C}
    \item{\code{temperature_opt}}{the mean temperature of the month, where NDVI is highest in °C}
    \item{\code{NDVI}}{A raster(stack) of NDVI values}
    \item{\code{FAPAR_max}}{constant, see Wang et al. (2017)}
    \item{\code{FAPAR_min}}{constant, see Wang et al. (2017)}
  }}
}
\value{
a RasterStack of NPP in g-C m^-2 d^-1 with each layer representing one day.
}
\description{
Calculate the net primary productivity (NPP) using the
 Carnegie-Ames-Stanford-Approach (CASA).
}
\references{
\insertRef{Potter1993}{remotepp} \cr
}
