% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_raster.R
\name{interpolate_raster}
\alias{interpolate_raster}
\title{interpolate raster}
\usage{
interpolate_raster(
  input_files,
  in_dates,
  out_dates,
  out_dir,
  interpolation_function = function(x, y, xout) (approx(x, y, xout))$y,
  interpolation_arguments = list(),
  overwrite = FALSE
)
}
\description{
Interpolate a raster timeseries from file to file.
}
